made the remove-sessions.sh work as a symlink. (so it can be symlinked from cron.daily)

+++ script/remove-sessions.sh
@@ -22,7 +22,12 @@ fi
 # Use the current directory for the RAILS_ROOT if it was not set
 if [ -z $RAILS_ROOT ]
 then
-    RAILS_ROOT=`pwd`
+    RAILS_ROOT=$(dirname $(readlink -f $0))/..
+fi
+
+if [ ! -d $RAILS_ROOT ]; then
+    echo "RAILS_ROOT does not exist."
+    exit 1
 fi
 
 find $RAILS_ROOT/tmp/sessions -mmin +$TIME -iname 'ruby_sess.*' -exec rm -v {} \;
